+++ CHANGELOG.md
@@ -6,6 +6,7 @@ omnibus-gitlab repository.
 10.4.0
 
 - Upgrade Ruby version to 2.3.6
+- Add warning to LoggingHelper
 
 10.3.0
 
@@ -32,6 +33,7 @@ omnibus-gitlab repository.
 - Bump bundled git version to 2.14.3 a2b4bedf
 - Render gitaly-ruby memory settings
 - Update pgbouncer recipe to better handle initial configuration
+- Add warning to LoggingHelper
 
 10.2.3